Gitolite permission acl
First dump of code with partial support of Gitolite acls. Access is granted if the user, or a group he/she is member of, listed in either the big conf or split conf acl.
There is NO CHECK of the actual permission (R/W) nor are the refs taken into account!